Add back in the frc simulation plugins for gazebo.
This includes gradle build rules and some basic unit tests.

This also includes the models and worlds originally
including in frcsim.

This also removes *.sdf from the .gitignore file.  If it is
still an issue in Visual C++, I suggest we add it back in
with a more targetted filter.